[NO-JIRA] Refactor email helpers
