+++ server/api/actions/users/create.js
@@ -1,5 +1,6 @@
 import loginUser from '../auth/private/login';
 import { datumCreate, datumLoad } from '../datum';
+import sendEmailVerificationEmail from '../auth/private/sendEmailVerificationEmail';
 
 import usersAllowedAttributes from './private/usersAllowedAttributes';
 
@@ -7,7 +8,6 @@ import lockPromise from 'utils/lock';
 import authentication from 'utils/authentication';
 import { hash } from 'utils/hash';
 import { find, emailFingerprint } from 'utils/find';
-import { sendEmailVerificationEmail } from 'utils/emailHelpers';
 import { UNAUTHORISED_WRITE } from 'utils/errorConstants';
 import reqSecure from 'utils/reqSecure';
 
@@ -48,9 +48,10 @@ export default function create(req) {
                   req.body.emailVerified = false;
                   datumCreate({ redisKey: 'users', user }, req).then(
                     newUser => {
-                      sendEmailVerificationEmail(newUser);
-                      loginUser(newUser);
-                      resolve({ message: 'User created' });
+                      sendEmailVerificationEmail(newUser).then(() => {
+                        loginUser(newUser);
+                        resolve({ message: 'User created' });
+                      });
                     },
                   );
                 }
